+++ app/storage/telegram/telegram_bot_storage.py
@@ -37,7 +37,13 @@ class TelegramBotStorage(Storage):
             message_id=result.message_id,
         )
 
-    async def get_file(self, file_id: str):
+    async def get_file(
+        self,
+        file_id: str,
+        *,
+        chat_id: str | None = None,
+        message_id: int | None = None,
+    ):
         await self._acquire()
         return await get_bot().get_file(file_id)